[None][fix] Match signedness in V2 probe bounds check
TokenSpan::size() returns int, while the block boundary uses size_t.
Cast the span size explicitly to preserve the existing comparison and
avoid -Werror=sign-compare failures in x86_64 and SBSA CI builds.

Validation: compiled the affected translation unit with -fsyntax-only,
-Wall, -Werror and -Wsign-compare in the shared B200 development container.
The original source reproduces the CI error; the corrected source passes.

+++ cpp/tensorrt_llm/batch_manager/kv_cache_manager_v2/kvCacheManager.cpp
@@ -237,7 +237,7 @@ std::optional<BlockKey> KvCacheManager::probeFirstNewBlockKey(
     int const blockSize = tokensPerBlock();
     int const blockIndex = match.numTokens / blockSize;
     size_t const begin = static_cast<size_t>(blockIndex) * static_cast<size_t>(blockSize);
-    if (begin + static_cast<size_t>(blockSize) > inputTokens.size())
+    if (begin + static_cast<size_t>(blockSize) > static_cast<size_t>(inputTokens.size()))
     {
         return std::nullopt;
     }
